Corrigido bug ao executar o seguinte código: telatxt = 1 classe a func iniclasse telatxt tela tela.msg(a:x + "\n") # Aqui o valor de a:x se perdia func x uint32 n0 n0 = 60000 ret n0 # Aqui n0 ainda Ã© 60000

+++ variavel.cpp
@@ -461,7 +461,7 @@ void TVariavel::MoverEnd(void * destino, TClasse * classe, TObjeto * objeto)
     case Instr::cUInt32:
         if (vetor <= 1)
         {
-            int x = *(short*)endvar;
+            int x = *(int*)endvar;
             *(int*)destino = x;
         }
         else
